+++ packages/datasource-mongoose/test/integration/review/collection_list.test.ts
@@ -5,6 +5,7 @@ import {
   ConditionTreeLeaf,
   Operator,
   Page,
+  PaginatedFilter,
   Projection,
   Sort,
 } from '@forestadmin/datasource-toolkit';
@@ -88,6 +89,34 @@ describe('MongooseCollection', () => {
       );
 
       expect(records).toEqual([{ message: 'C' }, { message: 'B' }, { message: 'A' }]);
+    });
+  });
+
+  describe('page & sort', () => {
+    describe('when a user wants to sort records on all pages', () => {
+      it('sort must be applied on all pages', async () => {
+        connection = await setupReview('collection_review_list');
+        const dataSource = new MongooseDatasource(connection);
+        const review = dataSource.getCollection('review');
+        await review.create(factories.caller.build(), [
+          { message: 'A' },
+          { message: 'B' },
+          { message: 'C' },
+          { message: 'D' },
+        ]);
+
+        const filter = new PaginatedFilter({
+          page: new Page(0, 2),
+          sort: new Sort({ field: 'message', ascending: false }),
+        });
+        const records = await review.list(
+          factories.caller.build(),
+          filter,
+          new Projection('message'),
+        );
+
+        expect(records).toEqual([{ message: 'D' }, { message: 'C' }]);
+      });
     });
   });