+++ packages/plugin-aws-s3/src/field/make-field-filterable.ts
@@ -1,12 +1,13 @@
 import type { Configuration } from '../types';
 import type { CollectionCustomizer } from '@forestadmin/datasource-customizer';
-import type { ColumnSchema } from '@forestadmin/datasource-toolkit';
+
+import { SchemaUtils } from '@forestadmin/datasource-toolkit';
 
 export default function makeFieldFilterable(
   collection: CollectionCustomizer,
   config: Configuration,
 ): void {
-  const schema = collection.schema.fields[config.sourcename] as ColumnSchema;
+  const schema = SchemaUtils.getColumn(collection.schema, config.sourcename, collection.name);
 
   for (const operator of schema.filterOperators) {
     collection.replaceFieldOperator(config.filename, operator, value => ({