+++ packages/_example/CHANGELOG.md
@@ -1,3 +1,51 @@
+# example [1.11.0](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.10.4...example@1.11.0) (2024-10-07)
+
+
+### Features
+
+* add pages in action forms ([#1179](https://github.com/ForestAdmin/agent-nodejs/issues/1179)) ([0b2dc0a](https://github.com/ForestAdmin/agent-nodejs/commit/0b2dc0a0d20b0f762aeb9eb6674693e05fd386c2))
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.49.0
+* **@forestadmin/datasource-dummy:** upgraded to 1.1.33
+* **@forestadmin/datasource-mongo:** upgraded to 1.3.23
+* **@forestadmin/datasource-mongoose:** upgraded to 1.8.3
+* **@forestadmin/datasource-sequelize:** upgraded to 1.10.2
+* **@forestadmin/datasource-sql:** upgraded to 1.14.28
+* **@forestadmin/datasource-toolkit:** upgraded to 1.42.0
+
+## example [1.10.4](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.10.3...example@1.10.4) (2024-10-04)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.48.0
+* **@forestadmin/datasource-dummy:** upgraded to 1.1.32
+* **@forestadmin/datasource-mongo:** upgraded to 1.3.22
+* **@forestadmin/datasource-mongoose:** upgraded to 1.8.2
+* **@forestadmin/datasource-sequelize:** upgraded to 1.10.1
+* **@forestadmin/datasource-sql:** upgraded to 1.14.27
+* **@forestadmin/datasource-toolkit:** upgraded to 1.41.0
+
+## example [1.10.3](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.10.2...example@1.10.3) (2024-10-04)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-mongo:** upgraded to 1.3.21
+* **@forestadmin/datasource-mongoose:** upgraded to 1.8.1
+
 ## example [1.10.2](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.10.1...example@1.10.2) (2024-09-27)